requirements.txt
flask
flask-sock
twilio
python-dotenv
pydub
websockets
soundfile
numpy